Don't depend on NVM_HOME being set
When using NVM-noinstall, for example, NVM_HOME may not be set. Even
with the installer, depending on the Windows environment, NVM_HOME
being set may require "restarting cmd" (or rather, reloading the
system env vars).

Instead of depending on this env var, assume the settings file lives
in the same dir as the executable. This is still hacky, and probably
won't work elsewhere.